consumer/requirements.txt
python-dotenv==1.0.0
pika==1.3.2
kafka-python==2.0.2
influxdb-client==1.36.1